# docs/swagger/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Convo API",
    "version": "1.0.0",
    "description": "\n      **Convo API Documentation**\n      \n      A comprehensive API for managing conversational forms, workspaces, and user interactions.\n      \n      ## Authentication\n      This API uses Clerk for authentication. Include your session token in the Authorization header:\n      `Authorization: Bearer <your-token>`\n      \n      ## Rate Limiting\n      API requests are rate limited per user:\n      - **Free Plan**: 100 requests/hour\n      - **Pro Plan**: 1000 requests/hour\n      - **Enterprise**: Custom limits\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Time when the limit resets (Unix timestamp)\n    ",
    "contact": {
      "name": "API Support",
      "email": "support@convo.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-generated-at": "2025-07-01T10:52:11.739Z",
    "x-generator": "convo-forms-swagger-generator"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    },
    {
      "url": "https://convo.ai/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token for user authentication"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "message": {
                  "type": "string",
                  "example": "Authentication required to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "AUTH_REQUIRED"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Access forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                },
                "message": {
                  "type": "string",
                  "example": "Insufficient permissions to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not Found"
                },
                "message": {
                  "type": "string",
                  "example": "The requested resource was not found"
                },
                "code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Validation Error"
                },
                "message": {
                  "type": "string",
                  "example": "Request validation failed"
                },
                "code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string",
                        "example": "email"
                      },
                      "message": {
                        "type": "string",
                        "example": "Invalid email format"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal Server Error"
                },
                "message": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "paths": {
...
    "/api/templates/{id}": {
      "get": {
        "summary": "Get a specific template",
        "description": "Retrieves detailed information about a specific template.\nReturns global templates or workspace templates the user has access to.\n",
        "tags": ["Templates"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Template retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Template"
                },
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Job Application Form",
                      "description": "Comprehensive job application form template",
                      "category": "HR",
                      "isGlobal": true,
                      "usageCount": 145,
                      "cloneCount": 23,
                      "formSchema": {
                        "fields": [
                          {
                            "id": "full_name",
                            "type": "text",
                            "label": "Full Name",
                            "required": true
                          }
                        ]
                      },
                      "createdAt": "2024-01-15T10:30:00Z",
                      "updatedAt": "2024-01-20T14:45:00Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update a specific template",
        "description": "Updates template properties such as name, description, form schema, and category.\nOnly workspace templates can be updated, and only by users with edit_template permission.\nGlobal templates cannot be modified by users.\n",
        "tags": ["Templates"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "Template name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Template description"
                  },
                  "formSchema": {
                    "type": "object",
                    "description": "Form configuration schema"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "HR",
                      "Marketing",
                      "Support",
                      "Sales",
                      "Education",
                      "Healthcare",
                      "Other"
                    ],
                    "description": "Business category"
                  },
                  "thumbnailUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Template preview image"
                  }
                }
              },
              "examples": {
                "updateTemplate": {
                  "summary": "Update template information",
                  "value": {
                    "name": "Updated Customer Survey",
                    "description": "Updated description for customer feedback template",
                    "category": "Marketing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "template": {
                      "$ref": "#/components/schemas/Template"
                    },
                    "message": {
                      "type": "string",
                      "example": "Template updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific template",
        "description": "Permanently deletes a template. Only workspace templates can be deleted,\nand only by users with edit_template permission. Global templates cannot be deleted.\n\n⚠️ Warning: This action cannot be undone. Forms created from this template will not be affected.\n",
        "tags": ["Templates"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Template deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}